+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project from 1.0.0 forward will be documented in thi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.7] - 2023-03-22
+
+### Added
+
+-   `pdo.getVersion()` return promise version of database
+-   **(DEV)** please ensure your driver is updated for lupdo version 3.1.7 and implements abstract `pdoDriver.getServerVersion()`.
+
 ## [3.1.6] - 2023-03-12
 
 ### Fixed